Moved `ecg_fixpeaks()` to `signal_fixpeaks()`.
