+++ neurokit2/ecg/__init__.py
@@ -3,7 +3,6 @@
 from .ecg_simulate import ecg_simulate
 from .ecg_clean import ecg_clean
 from .ecg_findpeaks import ecg_findpeaks
-from .ecg_fixpeaks import ecg_fixpeaks
 from .ecg_peaks import ecg_peaks
 from .ecg_segment import ecg_segment
 from .ecg_process import ecg_process
