+++ neurokit2/ecg/ecg_simulate.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import pandas as pd
 import numpy as np
 import scipy
 import math
